Multi-mode DRT possible
